core/pyproject.toml
[tool.poetry]
name = "covidflow-core"
version = "0.1.0"
description = "Rasa Stack/Core Server for covidflow"
authors = ["Nu Echo"]

[tool.poetry.dependencies]
rasa-integration-testing = "0.1.3"
python = "^3.6"
structlog = "^20.1.0"
python-json-logger = "^0.1.11"
pytz = "^2020.1"
analytics-python = "^1.2.9"

[tool.poetry.dev-dependencies]
pytest = "5.3.4"
pytest-cov = "2.8.1"
black = "19.10b0"
flake8 = "3.7.9"
isort = "^4.3.21"
mypy = "0.770"
pep8-naming = "0.10.0"
pylama = "^7.7.1"
"ruamel.yaml" = "^0.16.10"
sanic = "^19.12.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
